+++ seohyeon/ch6_12/src/main/java/core/mvc/Controller.java
@@ -1,4 +1,4 @@
-package next.controller;
+package core.mvc;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
