+++ seohyeon/ch6_12/src/main/java/core/mvc/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package next.controller;
+package core.mvc;
 
 import java.io.IOException;
 import javax.servlet.RequestDispatcher;
