+++ seohyeon/ch6_12/src/main/java/next/controller/HomeController.java
@@ -1,5 +1,6 @@
 package next.controller;
 
+import core.mvc.Controller;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
