+++ seohyeon/ch6_12/src/main/java/next/controller/LogoutController.java
@@ -1,9 +1,11 @@
 package next.controller;
 
+import core.mvc.Controller;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import next.util.UserSessionUtils;
 
 @WebServlet("/users/logout")
 public class LogoutController implements Controller {
