+++ seohyeon/ch6_12/src/main/java/next/controller/UpdateUserController.java
@@ -1,9 +1,11 @@
 package next.controller;
 
+import core.mvc.Controller;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import next.util.UserSessionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
